Broadcast widget update from TunnelManager
