+++ app/src/main/java/com/wireguard/android/model/TunnelManager.kt
@@ -5,10 +5,13 @@
  */
 package com.wireguard.android.model
 
+import android.appwidget.AppWidgetManager
 import android.content.BroadcastReceiver
+import android.content.ComponentName
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
+import android.os.Handler
 import androidx.databinding.BaseObservable
 import androidx.databinding.Bindable
 import com.wireguard.android.BR
@@ -18,6 +21,7 @@ import com.wireguard.android.backend.Backend
 import com.wireguard.android.configStore.ConfigStore
 import com.wireguard.android.di.getInjector
 import com.wireguard.android.model.Tunnel.Statistics
+import com.wireguard.android.providers.OneTapWidget
 import com.wireguard.android.util.ApplicationPreferences
 import com.wireguard.android.util.AsyncWorker
 import com.wireguard.android.util.ExceptionLoggers
@@ -38,6 +42,7 @@ class TunnelManager @Inject constructor(
     private val backend: Backend,
     private val context: Context,
     private val configStore: ConfigStore,
+    private val handler: Handler,
     private val prefs: ApplicationPreferences
 ) : BaseObservable() {
 
@@ -274,6 +279,13 @@ class TunnelManager @Inject constructor(
                 setLastUsedTunnel(tunnel)
             }
             saveState()
+            handler.postDelayed({
+                context.sendBroadcast(Intent(context, OneTapWidget::class.java).apply {
+                    val ids = AppWidgetManager.getInstance(context).getAppWidgetIds(ComponentName(context, OneTapWidget::class.java))
+                    putExtra(AppWidgetManager.EXTRA_APPWIDGET_IDS, ids)
+                    action = AppWidgetManager.ACTION_APPWIDGET_UPDATE
+                })
+            }, 1000L)
         }
     }
